[update][manipulation routines] Refine functions `reshape`, `resize`, `flatten`, `ravel` (#158)
I refined and fixed the functions `reshape` `resize` and methods
`flatten` 'ravel` so that:

1. Refine the function `reshape` so that it is working on any dimensions
and is working on both row-major and col-major. This also allows us to
change the order with the code:
```mojo
A = nm.random.randn(Shape(2,3,4))
A.reshape(A.shape, "F"))
```
2. Add a new method `resize` that reshapes the array in-place. `reshape`
method returns a copy of the array.
3. Refine the function `ravel` so that it allows to flatten the array by
either row-major or col-major.
4. Update the method `flatten` which has the same behavior as `ravel`.
5. And two auxiliary functions that generate a list of integers given
range, e.g., `[0,1,2,3]`.
6. Show order of the array during printing.

The functions are tested against 3darrays and against both "C" and "F"
orders.