[shape] Update `NDArrayShape`, making `size` a method (#181)
This PR makes some updates to the `NDArrayShape` and `NDArrayStrides`.

1. Make `NDArrayShape` and `NDArrayStrides` `@register_passable` instead
of `@register_passable("trivial")`.

Reason: This is to prevent implicit copy by reference. For example,
before change, the following code is not what we want and is prone to
mistakes.

```mojo
>>>var a = nm.NDArrayShape(1, 2, 3)
>>>var b = a
>>>b[0] = 10
>>>print(a)
# Shape: [10, 2, 3]
# Not what we want
```

After change, this is avoided.

```mojo
>>>var a = nm.NDArrayShape(1, 2, 3)
>>>var b = a
>>>b[0] = 10
>>>print(a)
# Shape: [1, 2, 3]
# It is better now
```

2. Change the `size` property into `size_of_array` method for
`NDArrayShape`.

Reason: The size information is only used when we want to know the size
of array from the shape. If it is a property, it must be calculated at
the initialization of the shape, causing unnecessary costs. Moreover, it
makes it difficult when we want to change the ith-value of the `Shape`,
or making a `Shape` object by joining two `shapes`, because the size is
not automatically updated.

By using `size_of_array` method, we ensure that the value is calculated
only when it is needed.