[ndarray][nditer] Add `NDIter` struct to iterate items in `NDArray` + add `copy` method (#188)
[ndarray][nditer] Add `NDIter` struct to iterate items in `NDArray`

## Add `_NDIter` struct

This struct is an iterator over items of `NDArray` according to the
memory layout of the array or a manually input order.

## Add `nditer` method

This method creates an instance of `NDIter` iterator. There are two
overloads. If no `order` argument is passed in, the iterator walks
through the array according to its memory layout. If `order` argument is
passed in, the iterator walks through the array according to this
`order`.

## Example

See the following example code and note the difference.

```mojo
fn main() raises:
    var a = nm.array[i8]("[[1, 2], [3, 4], [5, 6]]")
    print(a)
    for i in a.nditer():
        print(i, end=" ")
    print()
    for i in a.nditer("C"):
        print(i, end=" ")
    print()
    for i in a.nditer("F"):
        print(i, end=" ")
    print()

    a = a.reshape(Shape(3, 2), order="F")
    print(a)
    for i in a.nditer():
        print(i, end=" ")
    print()
    for i in a.nditer("C"):
        print(i, end=" ")
    print()
    for i in a.nditer("F"):
        print(i, end=" ")
    print()
```

```console
# C-order array
[[      1       2       ]
 [      3       4       ]
 [      5       6       ]]
2-D array  (3,2)  DType: int8  C-cont: True  F-cont: False  own data: True
1 2 3 4 5 6  # default order
1 2 3 4 5 6  # C order
1 3 5 2 4 6  # F order

# F-order array
[[      1       2       ]
 [      3       4       ]
 [      5       6       ]]
2-D array  (3,2)  DType: int8  C-cont: False  F-cont: True  own data: True
1 3 5 2 4 6  # default order
1 2 3 4 5 6  # C order
1 3 5 2 4 6  # F order
```

## Add a copy method

It is at the moment the same as `copyinit`, but will have different
behavior when we have view of array.