[core][print] Significantly increase the speed of printing arrays (#215)
This PR changes the approach in determining the min and max values of
the printable regions of an array. This significantly improves the speed
of printing arrays.

This improvement is particularly significant when we encounter very
large arrays. The speed increase can be x100000. See the following
comparison on a (10000, 1000) array.

```console
# before the change
2D-array  Shape(10000,1000)  Strides(1000,1)  DType: f64  C-cont: True  F-cont: False  own data: True
Time to print array: 19.190531999978703

# after the change
2D-array  Shape(10000,1000)  Strides(1000,1)  DType: f64  C-cont: True  F-cont: False  own data: True
Time to print array: 0.0001010000123642385
```