[fix][math] Fix error in Division of Array and a Scalar (#244)
# Error description
```mojo
fn main() raises:
    var list: List[Scalar[nm.f64]] = List[Scalar[nm.f64]]()
    for i in range(200):
        list.append(Scalar[nm.f64](1e-5))

    var array = nm.array[nm.f64](list).reshape(nm.Shape(200, 1))
    var den = nm.exp(-array) + 1.0
    var result = Scalar[nm.f64](1.0) / (den)
    print("r: ", r)
```
The above code produces an array filled with 2.0 which is wrong since
the expected result is an array filled with 0.5. This error occurred due
to the symmetric implementation of all arithmetic operators i.e `array
+-*/ scalar`, but this results in wrong value for division since it
always calculates `array / scalar` even when the user input is (scalar /
array).

# Solution
This PR fixes this error by adding a new function to the backends that
calculate the array and scalar arithmetic operations so that array /
scalar and scalar / array result in the correct values.