[core][Matrix] Return Matrix views.  (#280)
This PR lays the groundwork for returning views in both the Matrix and
NDArray types.
It also implements full support for Matrix views. The design for
returning view is described in the NuMojo enhancement proposal #279

Support for NDArray views will be completed in a subsequent PR.

Currently test workflow is commented out. Since Mojo has removed the
`test` command, we will use the `TestSuite` to run tests hereafter. The
tests will be fixed in the follow up PR.

---------

Co-authored-by: ZHU Yuhao 朱宇浩 <dr.yuhao.zhu@outlook.com>